- putting components into different files - changing to appropriate file names
